config: Disable zswap compressed cache by default
Per GitHub issue [1], zswap causes stutters when system is under memory
pressure.  This is completely undesirable behavior for desktops, but
might be tolerable in a server environment where interactivity and
latency is not an issue.

Turn off since we prefer less jank at the cost of total theoretical
memory capacity.

[1] https://github.com/zen-kernel/zen-kernel/issues/356